+++ Agp2p.Web/admin/repayment/repay_earlier.aspx.cs
@@ -177,7 +177,7 @@ namespace Agp2p.Web.admin.repayment
             public string RepayTime { get; set; }//实还时间
             public int DayCount { get; set; }//提前天数
             public int Category { get; set; }//产品
-            public int ProfitRate { get; set; }//年化利率
+            public decimal ProfitRate { get; set; }//年化利率
             public string RepaymentType { get; set; }//年化利率
             public int ProjectStatus { get; set; }//项目状态
             public int RepayStatus { get; set; }//还款状态
