+++ Agp2p.Web/admin/repayment/repay_manage.aspx.cs
@@ -164,7 +164,7 @@ namespace Agp2p.Web.admin.repayment
             public string ShouldRepayTime { get; set; }//应还时间
             public string RepayTime { get; set; }//实还时间
             public int Category { get; set; }//产品
-            public int ProfitRate { get; set; }//年化利率
+            public decimal ProfitRate { get; set; }//年化利率
             public string RepaymentType { get; set; }//年化利率
             public int ProjectStatus { get; set; }//项目状态
             public int RepayStatus { get; set; }//还款状态
